backend/requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.11.0
djoser==2.1.0